moved mongo_client to mongo package
