+++ service/user_service.py
@@ -1,7 +1,6 @@
 import logging
 
-from domain.User import User
-from persistence.mongo_client import add_aliases
+from persistence.mongo.mongo_client import add_aliases
 
 logger = logging.Logger(__name__)
